refactored BountyClosed dir structure
