+++ components/BountyClosed/index.js
